chore: refactor for remix
